minor refactoring of feature encoders module
