+++ easyflow/feature_encoders/__init__.py
@@ -1,7 +1,8 @@
 """Keras and tensorflow feature column pipeline
 """
 
-from .feature_encoder import CategoricalFeatureEncoder, EmbeddingFeatureEncoder, NumericalFeatureEncoder, get_unique_vocab
+from .categorical_encoders import CategoricalFeatureEncoder, EmbeddingFeatureEncoder
+from .numerical_encoders import NumericalFeatureEncoder
 from .transformer import FeatureColumnTransformer, FeatureUnionTransformer
